refactor(inpoint): extract flv_chunker test module to sibling file
